Exception handling + controllers mod + checkOwnership method

+++ tem/src/main/java/com/java/tem/controller/DatiCorsaController.java
@@ -1,21 +1,25 @@
 package com.java.tem.controller;
 
+import com.java.tem.exceptions.DoesNotBelongToAzienda;
 import com.java.tem.model.accountservice.entity.AccountService;
 import com.java.tem.model.accountservice.entity.Utente;
 import com.java.tem.model.programmacorseservice.entity.daticorsaservice.DatiCorsa;
 import com.java.tem.model.programmacorseservice.entity.daticorsaservice.DatiCorsaService;
 import java.util.List;
 import javax.validation.Valid;
+import org.dom4j.rule.Mode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.ModelAndView;
 
 @Controller
 public class DatiCorsaController {
@@ -34,50 +38,69 @@ public class DatiCorsaController {
   }
 
   @GetMapping("/daticorsa/edit/{id}")
-  public String showFormDatiCorsa(@PathVariable("id") Long id, Model model) {
+  public ModelAndView showFormDatiCorsa(@PathVariable("id") Long id, Model model) {
     DatiCorsa datiCorsa = this.datiCorsaService.getDatiCorsa(id)
-        .orElseThrow(() -> new IllegalArgumentException("Invalid Conducente Id:" + id));
+        .orElseThrow(() -> new IllegalArgumentException("Invalid Dati Corsa Id:" + id));
     model.addAttribute("datiCorsa", datiCorsa);
-    return "edit-daticorsa";
+
+    Utente utente = accountService.getLoggedUser();
+    try {
+      if (!datiCorsaService.checkOwnership(datiCorsa, utente)) {
+        throw new DoesNotBelongToAzienda("I dati corsa non appartengono alla tua azienda");
+      }
+      return new ModelAndView("edit-daticorsa");
+    } catch (DoesNotBelongToAzienda exc) {
+      model.addAttribute("error", exc.getMessage());
+      return new ModelAndView("redirect:/home", (ModelMap) model);
+    }
   }
 
   @PostMapping("/daticorsa/submit")
-  public String processDatiCorsa(@ModelAttribute("datiCorsa") @Valid DatiCorsa datiCorsa,
+  public ModelAndView processDatiCorsa(@ModelAttribute("datiCorsa") @Valid DatiCorsa datiCorsa,
                                  BindingResult bindingResult,
                                  Model model) {
     if (bindingResult.hasErrors()) {
-      return "insert-daticorsa";
+      return new ModelAndView("insert-daticorsa");
     }
     Utente utente = accountService.getLoggedUser();
     datiCorsa.setAzienda(utente);
     this.datiCorsaService.addDatiCorsa(datiCorsa);
     model.addAttribute(datiCorsa);
-    return "dati-corsa-success";
+    return new ModelAndView("dati-corsa-success");
+
+
+
   }
 
   @PostMapping("daticorsa/update/{id}")
-  public String updateDatiCorsa(@PathVariable("id") Long id, DatiCorsa datiCorsa,
-                                BindingResult result,
-                                Model model) {
+  public ModelAndView updateDatiCorsa(@PathVariable("id") Long id, DatiCorsa datiCorsa,
+                              BindingResult result,
+                              Model model) {
     if (result.hasErrors()) {
-      return "edit-daticorsa";
+      return new ModelAndView("update-daticorsa");
     }
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    String currentUserName = authentication.getName();
-    Utente utente = accountService.getUserByUsername(currentUserName);
+    Utente utente = accountService.getLoggedUser();
     datiCorsa.setAzienda(utente);
     this.datiCorsaService.updateDatiCorsa(datiCorsa);
-    return "update-success";
+    return new ModelAndView("update-success");
   }
 
   @GetMapping("daticorsa/delete/{id}")
-  public String deleteDatiCorsa(@PathVariable("id") Long id, Model model)
+  public ModelAndView deleteDatiCorsa(@PathVariable("id") Long id, Model model)
       throws IllegalArgumentException {
     DatiCorsa datiCorsa = datiCorsaService.getDatiCorsa(id)
         .orElseThrow(() -> new IllegalArgumentException("Invalid daticorsa Id:" + id));
-    datiCorsaService.deleteDatiCorsa(datiCorsa);
+    Utente utente = accountService.getLoggedUser();
 
-    return "home";
+    try {
+      if (!datiCorsaService.checkOwnership(datiCorsa, utente)) {
+        throw new DoesNotBelongToAzienda("I dati corsa non appartengono alla tua azienda");
+      }
+      datiCorsaService.deleteDatiCorsa(datiCorsa);
+    } catch (DoesNotBelongToAzienda exc) {
+      model.addAttribute("error", exc.getMessage());
+    }
+    return new ModelAndView("redirect:/home", (ModelMap) model);
   }
 
   @GetMapping("/daticorsa/list")
@@ -91,7 +114,7 @@ public class DatiCorsaController {
   }
 
   @GetMapping("/daticorsa/list/{aziendaId}")
-  public String listDatiCorsa(@PathVariable("aziendaId") Long id, Model model) {
+  public String listDatiCorsaByAzienda(@PathVariable("aziendaId") Long id, Model model) {
     Utente azienda = accountService.getUserById(id);
     boolean isAdmin = accountService.isAdmin();
